requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
Pillow
djangorestframework-simplejwt
drf-yasg
django-filter
celery
redis
django-cors-headers
